Add support for enqueueing groups
Frontend + backend changes to support students enqueueing as groups.
If the student belongs to a group in their current course it will
automatically be attached to the request. From there, it is up to the
instructor to toggle "group mode," for the queue. "Group mode" restricts
to one request per group as opposed to per user.

The requester will still be attached, (i.e. you'll still see the group
member who enqueued's name on the request) however anyone in the same
group will be able to edit/cancel the request.

These groups will be created in the admin interface, (not editable to
students to prevent abuse) which has not been created yet.