Add heartbeat controller
After a brief outage caused by a letsencrypt renewal failure, it's
probably a good idea to keep an eye on this since a fair number of
courses depend on it now.

Add a page that can be queried in a script without the need to
authenticate to determine at least that the rails app is functional.