Update queue pop button functionality
to allow popping people from the queue that are pinned by others.

The pop order is now: pinned by you, not pinned, pinned by others, which
mirrors the way it is displayed in the UI.

Since the queue pop button can now pop up to $requests.length requests,
this closes #98.