Add exclusive mode
Exclusive mode prevents users who are not on the group list from
enqueueing requests. Group mode and exclusive mode are not mutually
exclusive. If you have group mode off but exclusive mode on, we will
simply check that a user is in _a_ group, not whether or not their group
already has an existing request.