broadcast_request_change: add optional queue param
In #164 we started serializing outstanding requests and reordering the
frontend queue entries accordingly. Since the merge functionality
broadcast its own change, it didn't pick up this change + broke as the
frontend was expecting this 'queue' parameter. To make it easier to keep
all request changes in a certain shape, update broadcast_request_change
to accept an optional queue parameter and move the merge functionality
under there.