Implement popup dialog box for ping messages
A modal window with a textarea for text input will appear when the user
clicks the "Ping" link. The user can close the modal window by clicking
outside of it, on the close window icon, or the "Send Ping" button.
Currently the "Send Ping" button does correctly send the ping to the
desired student.