Optionally add jitter to initial requests
There have been many reports over the years of students who script
clicking of 'request help' so they land on the top of the queue when it
first opens. Those without are left at the bottom of the queue,
potentially waiting hours for help.

This adds an opt-in feature to scramble the request timestamp by +/- 10
seconds during the first minute a queue is open. It is:

* implemented entirely server-side, hopefully minimizing gameability
* automatic when enabled, so instructors can set it and forget it
* more-or-less transparent to students, though they may notice their
  queue position change during the first minute

Hopefully this is enough to remove the edge for those using scripts and
level the field. I marked the toggle as 'beta'; we can iterate on the
approach depending on what instructors find in practice.